runtests.py: don't run tests unless explicitly enabled
Previously, when --enabled was not given, all tests were run.  That was meant
to simplify running all tests, but that is rare and more likely to happen by
mistake when not selecting tests correctly in the automated testing, so now we
revert the logic and don't select a test if no --enable filter is set.