ObexTransportAgent.cpp: properly shut down connection (FDO #91485)
Apparently there's a race condition in the OBEX transport that causes the
connection to phones via Bluetooth to be shut down prematurely.  Some phones
react by doing a slow sync instead of an incremental sync the next time.

Waiting during shutdown should address the problem (however, it was not
possible to confirm this).