SoupTransportAgent: fix disabled SSL checking
The SoupTransportAgent modernization in 1.5.3 led to SSL checking
always being enabled because the default changed from disabled to
enabled and SyncEvolution did not set it.

Worse, in older versions it probably (untested) was always disabled
because it was not enabled either.

Now the checking of SSL is always set explicitly and thus always
mirrors the SyncEvolution configuration.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>